Minor edits to ssh syntax
The ssh syntax was inconsistent and confusing, sometimes using "ssh
git@server" and sometimes using "ssh gitolite".  This patch standardizes
on the conventions declared in the doc/0-INSTALL.mkd file.

After the concept of creating an ssh shortcut gets introduced in
doc/6-ssh-troubleshooting.mkd it is safe to use the other syntax.